third_party/WebKit/LayoutTests/http/tests/devtools/elements/insert-node-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that elements panel updates dom tree structure upon node insertion.


Running: testDumpInitial
========= Original ========
- <div id="container">
      <div id="child1"></div>
      <div id="child2"></div>
      <div id="child3"></div>
  </div>

Running: testInsertBefore
===== Inserted before =====
- <div id="container">
      <div id="child-before"></div>
      <div id="child1"></div>
      <div id="child2"></div>
      <div id="child3"></div>
  </div>

Running: testInsertMiddle
===== Inserted middle =====
- <div id="container">
      <div id="child-before"></div>
      <div id="child1"></div>
      <div id="child-middle"></div>
      <div id="child2"></div>
      <div id="child3"></div>
  </div>

Running: testAppend
======== Appended =========
- <div id="container">
      <div id="child-before"></div>
      <div id="child1"></div>
      <div id="child-middle"></div>
      <div id="child2"></div>
      <div id="child3"></div>
      <div id="child-after"></div>
  </div>

Running: testAppendWithText
======== Appended with text=========
- <div id="container">
      <div id="child-before"></div>
      <div id="child1"></div>
      <div id="child-middle"></div>
      <div id="child2"></div>
      <div id="child3"></div>
      <div id="child-after"></div>
      <div id="child-with-text" style="display: none;">Text</div>
  </div>
Success: child text is bound

Running: testInsertFirstTextNode
======== Inserted first text node =========
- <div id="container">
      <div id="child-before"></div>
      <div id="child1"></div>
      <div id="child-middle"></div>
      <div id="child2"></div>
      <div id="child3">First text</div>
      <div id="child-after"></div>
      <div id="child-with-text" style="display: none;">Text</div>
  </div>
Success: child text is bound
